+++ src/main/java/it/aldi/app/service/job_management/impl/JobManagementServiceImpl.java
@@ -1,8 +1,9 @@
-package it.aldi.app.service.manage_job.impl;
+package it.aldi.app.service.job_management.impl;
 
 import it.aldi.app.controller.rest.dto.request.AddJobCmd;
+import it.aldi.app.domain.Organization;
 import it.aldi.app.service.domain.JobService;
-import it.aldi.app.service.manage_job.JobManagementService;
+import it.aldi.app.service.job_management.JobManagementService;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -15,7 +16,7 @@ public class JobManagementServiceImpl implements JobManagementService {
     }
 
     @Override
-    public void addJob(AddJobCmd addJobCmd, Long organizationId) {
+    public void addJob(AddJobCmd cmd, Organization organization) {
         // TODO: implementation
     }
 }
