+++ src/main/java/it/aldi/app/service/subject_management/impl/SubjectManagementServiceImpl.java
@@ -1,4 +1,4 @@
-package it.aldi.app.service.manage_subject.impl;
+package it.aldi.app.service.subject_management.impl;
 
 import it.aldi.app.controller.rest.dto.request.AddSubjectCmd;
 import it.aldi.app.controller.rest.dto.response.SubjectDto;
@@ -7,20 +7,20 @@ import it.aldi.app.domain.Subject;
 import it.aldi.app.domain.SubjectType;
 import it.aldi.app.service.domain.SubjectService;
 import it.aldi.app.service.domain.SubjectTypeService;
-import it.aldi.app.service.manage_subject.ManageSubjectService;
+import it.aldi.app.service.subject_management.SubjectManagementService;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
-public class ManageSubjectServiceImpl implements ManageSubjectService {
+public class SubjectManagementServiceImpl implements SubjectManagementService {
 
     private final SubjectService subjectService;
 
     private final SubjectTypeService subjectTypeService;
 
-    public ManageSubjectServiceImpl(SubjectService subjectService,
-                                    SubjectTypeService subjectTypeService) {
+    public SubjectManagementServiceImpl(SubjectService subjectService,
+                                        SubjectTypeService subjectTypeService) {
         this.subjectService = subjectService;
         this.subjectTypeService = subjectTypeService;
     }
